Move and rename refactors for load dashboard handler
-  `dashboard` is root for all high level dashboard commands
-  `loadDashboardHandler` dir contains all logic for handling the LoadDashboard command
-  the `loadDashboardHandler` dir contains `index.ts` with the `loadDashboardHandler` saga
-  all other code and sub-sagas and generators related to loading dashboard are under the `loadDashboardHandler` directory

JIRA: RAIL-3383
