+++ libs/sdk-ui-dashboard/src/model/commandHandlers/dashboard/loadDashboardHandler/index.ts
@@ -1,31 +1,31 @@
 // (C) 2021 GoodData Corporation
 import { SagaIterator } from "redux-saga";
 import { all, call, put, SagaReturnType } from "redux-saga/effects";
-import { LoadDashboard } from "../../commands/dashboard";
-import { dispatchDashboardEvent } from "../../eventEmitter/eventDispatcher";
-import { dashboardLoaded } from "../../events/dashboard";
-import { filterContextActions } from "../../state/filterContext";
-import { insightsActions } from "../../state/insights";
-import { layoutActions } from "../../state/layout";
-import { loadingActions } from "../../state/loading";
-import { DashboardContext } from "../../types/commonTypes";
+import { LoadDashboard } from "../../../commands/dashboard";
+import { dispatchDashboardEvent } from "../../../eventEmitter/eventDispatcher";
+import { dashboardLoaded } from "../../../events/dashboard";
+import { filterContextActions } from "../../../state/filterContext";
+import { insightsActions } from "../../../state/insights";
+import { layoutActions } from "../../../state/layout";
+import { loadingActions } from "../../../state/loading";
+import { DashboardContext } from "../../../types/commonTypes";
 import { IDashboardWithReferences } from "@gooddata/sdk-backend-spi";
 import { resolveDashboardConfig } from "./resolveDashboardConfig";
-import { configActions } from "../../state/config";
-import { PromiseFnReturnType } from "../../types/sagas";
-import { dateFilterConfigActions } from "../../state/dateFilterConfig";
+import { configActions } from "../../../state/config";
+import { PromiseFnReturnType } from "../../../types/sagas";
+import { dateFilterConfigActions } from "../../../state/dateFilterConfig";
 import { DateFilterMergeResult, mergeDateFilterConfigWithOverrides } from "./mergeDateFilterConfigs";
 import { resolvePermissions } from "./resolvePermissions";
-import { permissionsActions } from "../../state/permissions";
+import { permissionsActions } from "../../../state/permissions";
 import { loadCatalog } from "./loadCatalog";
 import { loadDashboardAlerts } from "./loadDashboardAlerts";
-import { catalogActions } from "../../state/catalog/index";
-import { alertsActions } from "../../state/alerts/index";
+import { catalogActions } from "../../../state/catalog";
+import { alertsActions } from "../../../state/alerts";
 import { batchActions } from "redux-batched-actions";
-import { internalErrorOccurred } from "../../events/general";
+import { internalErrorOccurred } from "../../../events/general";
 import { loadUser } from "./loadUser";
-import { userActions } from "../../state/user";
-import { metaActions } from "../../state/meta";
+import { userActions } from "../../../state/user";
+import { metaActions } from "../../../state/meta";
 
 function loadDashboardFromBackend(ctx: DashboardContext): Promise<IDashboardWithReferences> {
     const { backend, workspace, dashboardRef } = ctx;
@@ -33,7 +33,7 @@ function loadDashboardFromBackend(ctx: DashboardContext): Promise<IDashboardWith
     return backend.workspace(workspace).dashboards().getDashboardWithReferences(dashboardRef!);
 }
 
-export function* loadDashboardCommandHandler(ctx: DashboardContext, cmd: LoadDashboard): SagaIterator<void> {
+export function* loadDashboardHandler(ctx: DashboardContext, cmd: LoadDashboard): SagaIterator<void> {
     try {
         yield put(loadingActions.setLoadingStart());
