+++ libs/sdk-ui-dashboard/src/model/commandHandlers/dashboard/loadDashboardHandler/loadCatalog.ts
@@ -1,5 +1,5 @@
 // (C) 2021 GoodData Corporation
-import { DashboardContext } from "../../types/commonTypes";
+import { DashboardContext } from "../../../types/commonTypes";
 import { IWorkspaceCatalog } from "@gooddata/sdk-backend-spi";
 
 export function loadCatalog(ctx: DashboardContext): Promise<IWorkspaceCatalog> {
