+++ libs/sdk-ui-dashboard/src/model/commandHandlers/dashboard/loadDashboardHandler/loadUser.ts
@@ -1,6 +1,6 @@
 // (C) 2021 GoodData Corporation
 import { IUser } from "@gooddata/sdk-backend-spi";
-import { DashboardContext } from "../../types/commonTypes";
+import { DashboardContext } from "../../../types/commonTypes";
 
 export function loadUser(ctx: DashboardContext): Promise<IUser> {
     const { backend } = ctx;
